correction

+++ get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: jdarcour <jdarcour@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 02:44:12 by jdarcour          #+#    #+#             */
-/*   Updated: 2022/11/30 22:38:07 by jdarcour         ###   ########.fr       */
+/*   Updated: 2022/12/02 15:23:37 by jdarcour         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
